misc: reorganize file directory
