+++ frontend/src/features/coordinate_upload/CoordinateFileUpload.tsx
