+++ frontend/src/features/coordinate_upload/ManualCoordinateUpload.tsx
