+++ frontend/src/features/coordinate_upload/YieldFileUpload.tsx
