+++ frontend/src/features/farm/index.ts
@@ -1,9 +1,9 @@
 export { default as FarmBiocharForm } from "./FarmBiocharForm";
-export { default as ManualCoordinateUpload } from "./ManualCoordinateUpload";
+export { default as ManualCoordinateUpload } from "../coordinate_upload/ManualCoordinateUpload";
 export { default as BiocharSettings } from "./BudgetSettings";
 export { default as FieldsList } from "./FieldsList";
 export type { FieldEntry } from "./FieldsList";
-export { default as FileUploadSection } from "./FileUploadSection";
+export { default as FileUploadSection } from "../coordinate_upload/FileUploadSection";
 export { default as SubmitSection } from "./SubmitSection";
-export { default as CoordinateFileUpload } from "./CoordinateFileUpload";
-export { default as YieldFileUpload } from "./YieldFileUpload";
+export { default as CoordinateFileUpload } from "../coordinate_upload/CoordinateFileUpload";
+export { default as YieldFileUpload } from "../coordinate_upload/YieldFileUpload";
